+++ Source/lib/embeddedLapack/src/LinearAlgebra/hankel.c
@@ -12,9 +12,9 @@
  * Argument step is how the shift of hankel matrix should be. Normaly, you want shift = 0.
  */
 
-void hankel(double* A, double* H, int length, int step) {
+void hankel(float* A, float* H, int length, int step) {
 
-	memset(H, 0, length*length*sizeof(double));
+	memset(H, 0, length*length*sizeof(float));
 
 	for(int i = 0; i < length; i++){
 		for(int j = 0; j < length; j++){
