+++ Source/lib/embeddedLapack/src/LinearAlgebra/mdiag.c
@@ -11,7 +11,7 @@
  * A size row x column and B size column x 1
  */
 
-void mdiag(double* A, double* B, int row, int column){
+void mdiag(float* A, float* B, int row, int column){
 
 	// This will take the diagonal of matrix a and turn it to a vector matrix
 	for(int i = 0; i < row; i++){
