+++ Source/lib/embeddedLapack/src/LinearAlgebra/pinv.c
@@ -10,13 +10,13 @@
  * Pseudo inverse of A with size row x column
  * Notice that pseudo inverse of A will return the size column x row if A is size row x column
  */
-void pinv(double* A, int row, int column) {
+void pinv(float* A, int row, int column) {
 	/*
 	 * Do SVD
 	 */
-	double U[row * row];
-	double S[row * column];
-	double V[column * column];
+	float U[row * row];
+	float S[row * column];
+	float V[column * column];
 	svd(A, U, S, V, row, column);
 
 	/*
@@ -28,11 +28,11 @@ void pinv(double* A, int row, int column) {
 	 */
 
 	if (row > column) {
-		double U_econ[row * column];
+		float U_econ[row * column];
 		cut(U, row, row, U_econ, 0, row - 1, 0, column - 1);
-		double S_econ[column * column];
+		float S_econ[column * column];
 		cut(S, row, column, S_econ, 0, column - 1, 0, column - 1);
-		double T[column * row];
+		float T[column * row];
 
 		// Inverse of S
 		inv(S_econ, column);
@@ -47,11 +47,11 @@ void pinv(double* A, int row, int column) {
 		mul(V, T, false, A, column, column, row);
 
 	} else {
-		double S_econ[row * row];
+		float S_econ[row * row];
 		cut(S, row, column, S_econ, 0, row - 1, 0, row - 1);
-		double V_econ[column * row];
+		float V_econ[column * row];
 		cut(V, column, column, V_econ, 0, column - 1, 0, row - 1);
-		double T[row * row];
+		float T[row * row];
 
 		// Inverse of S
 		inv(S_econ, row);
