+++ Source/lib/embeddedLapack/src/LinearAlgebra/power.c
@@ -9,10 +9,10 @@
 
 /*
  * Take the power values of every element of matrix A, size row x column
- * MATLAB code A.^n, where n is the double value
+ * MATLAB code A.^n, where n is the float value
  */
 
-void power(double* A, int row, int column, double value){
+void power(float* A, int row, int column, float value){
 	for(int i = 0; i < row; i++)
 		for(int j = 0; j < column; j++){
 			*A = powf(*(A), value);
